Progress made on feature test for Trips Dashboard

+++ app/models/trip.rb
@@ -21,7 +21,7 @@ class Trip < ActiveRecord::Base
   end
 
   def self.average_duration
-    Trip.average(:duration).to_f
+    Trip.average(:duration).to_i
   end
 
   def self.longest_ride
